refactor(webpack): move the async chunk layout plugin into template-webpack-plugin
The layout logic is DSL-agnostic (bundle names come from resolved modules,
the layer from module.layer), so export it as LynxAsyncChunkLayoutPlugin
next to the map it consumes. Other DSL plugins (e.g. TTML) can apply it
alongside LynxTemplatePlugin instead of duplicating it.
